requirements.txt
pytest~=8.2.0
selenium~=4.20.0
requests~=2.32.0
allure-python-commons~=2.13.5
data~=0.4